+++ types/src/monthUtils.d.ts
@@ -1,0 +1,3 @@
+export declare function getMonthsNames(): string[];
+export declare function getShortMonthsNames(): string[];
+export declare function getNarrowMonthsNames(): string[];