+++ types/src/weekUtils.d.ts
@@ -1,0 +1,3 @@
+export declare function getWeekDayssNames(): string[];
+export declare function getShortWeekDayssNames(): string[];
+export declare function getNarrowWeekDayssNames(): string[];