+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # NEXT VERSION
 
+- Feature: add "monthUtils" with functions to retrieve the months names in the user's locale
+- Feature: add "weekUtils" with functions to retrieve the week days' names in the user's locale
 - Feature: add temparature helper to obtain temprature units and convert them
 - Feature: add temperature formatter